Adds apython svgutils and SVG file maker
Call makeSVG from the command line with at least
the follow arguments:

-i (float)
-e (path/to/file.svg)+

and receive in the stdout a path to an SVG file.

Other options come after -o:

-o name (value)+

Including the argument

outDir (path/to/dir)

will change the output path.
Default is working directory.
Absolute paths are allowed.

Including the argument

size (int int)

will set the size of the viewBox,
or the visible part of the SVG file.